move DirEntryExt into util module
